Role Module

+++ app/Http/Controllers/PermissionController.php
@@ -22,37 +22,49 @@ class PermissionController extends BaseController
      * load menu index route
      */
     public function index(){
-        $this->setPageData('Permission', 'Permission', 'fas fa-th-list');
-        $data = $this->service->index();
-        return view('permission.index', compact('data'));
+        if(permission('permission-access')){
+            $this->setPageData('Permission', 'Permission', 'fas fa-th-list');
+            $data = $this->service->index();
+            return view('permission.index', compact('data'));
+        }else{
+            return $this->unauthorized_access_blocked();
+        }
     }
 
     /**
      * get datatable data
      */
     public function get_datatable_data(Request $request){
-        if($request->ajax()){
-            $output = $this->service->get_datatable_data($request);
+        if(permission('permission-access')){
+            if($request->ajax()){
+                $output = $this->service->get_datatable_data($request);
+            }else{
+                $output = ['status' => 'error', 'message' => 'Unauthorized access blocked!'];
+            }
+    
+            return response()->json($output);
         }else{
-            $output = ['status' => 'error', 'message' => 'Unauthorized access blocked!'];
+            return $this->unauthorized_access_blocked();
         }
-
-        return response()->json($output);
     }
 
     /**
      * store or update data
      */
     public function store(PermissionFormRequest $request){
-        if($request->ajax()){
-            $result = $this->service->store($request);
-            if($result){
-                return $this->response_json($status="success", $message="Data has been saved successfully!", $data=null, $response_code=200);
+        if(permission('permission-add')){
+            if($request->ajax()){
+                $result = $this->service->store($request);
+                if($result){
+                    return $this->response_json($status="success", $message="Data has been saved successfully!", $data=null, $response_code=200);
+                }else{
+                    return $this->response_json($status="error", $message="Data can't Save!", $data=null, $response_code=204);
+                }
             }else{
-                return $this->response_json($status="error", $message="Data can't Save!", $data=null, $response_code=204);
+                return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
             }
         }else{
-            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
+            return $this->unauthorized_access_blocked();
         }
     }
 
@@ -60,15 +72,19 @@ class PermissionController extends BaseController
      * edit menu data
      */
     public function edit(Request $request){
-        if($request->ajax()){
-            $data = $this->service->edit($request);
-            if($data->count() > 0){
-                return $this->response_json($status="success", $message=null, $data=$data, $response_code=201);
+        if(permission('permission-edit')){
+            if($request->ajax()){
+                $data = $this->service->edit($request);
+                if($data->count() > 0){
+                    return $this->response_json($status="success", $message=null, $data=$data, $response_code=201);
+                }else{
+                    return $this->response_json($status="error", $message="Data Not Found", $data=null, $response_code=204);
+                }
             }else{
-                return $this->response_json($status="error", $message="Data Not Found", $data=null, $response_code=204);
+                return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
             }
         }else{
-            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
+            return $this->unauthorized_access_blocked();
         }
     }
 
@@ -76,15 +92,19 @@ class PermissionController extends BaseController
      * update permission
      */
     public function update(PermissionUpdateRequest $request){
-        if($request->ajax()){
-            $result = $this->service->update($request);
-            if($result){
-                return $this->response_json($status="success", $message="Data has been updated successfully!", $data=null, $response_code=200);
+        if(permission('permission-edit')){
+            if($request->ajax()){
+                $result = $this->service->update($request);
+                if($result){
+                    return $this->response_json($status="success", $message="Data has been updated successfully!", $data=null, $response_code=200);
+                }else{
+                    return $this->response_json($status="error", $message="Data can't Update!", $data=null, $response_code=204);
+                }
             }else{
-                return $this->response_json($status="error", $message="Data can't Update!", $data=null, $response_code=204);
+                return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
             }
         }else{
-            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
+            return $this->unauthorized_access_blocked();
         }
     }
 
@@ -92,15 +112,19 @@ class PermissionController extends BaseController
      * delete menu data
      */
     public function delete(Request $request){
-        if($request->ajax()){
-            $result = $this->service->delete($request);
-            if($result){
-                return $this->response_json($status="success", $message="Data deleted successfully!", $data=null, $response_code=200);
+        if(permission('permission-delete')){
+            if($request->ajax()){
+                $result = $this->service->delete($request);
+                if($result){
+                    return $this->response_json($status="success", $message="Data deleted successfully!", $data=null, $response_code=200);
+                }else{
+                    return $this->response_json($status="error", $message="Failed to delete data!", $data=null, $response_code=204);
+                }
             }else{
-                return $this->response_json($status="error", $message="Failed to delete data!", $data=null, $response_code=204);
+                return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
             }
         }else{
-            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
+            return $this->unauthorized_access_blocked();
         }
     }
 
@@ -108,15 +132,19 @@ class PermissionController extends BaseController
      * bulk delete data from database
      */
     public function bulk_delete(Request $request){
-        if($request->ajax()){
-            $result = $this->service->bulk_delete($request);
-            if($result){
-                return $this->response_json($status="success", $message="Data deleted successfully!", $data=null, $response_code=200);
+        if(permission('permission-bulk-delete')){
+            if($request->ajax()){
+                $result = $this->service->bulk_delete($request);
+                if($result){
+                    return $this->response_json($status="success", $message="Data deleted successfully!", $data=null, $response_code=200);
+                }else{
+                    return $this->response_json($status="error", $message="Failed to delete data!", $data=null, $response_code=204);
+                }
             }else{
-                return $this->response_json($status="error", $message="Failed to delete data!", $data=null, $response_code=204);
+                return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
             }
         }else{
-            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
+            return $this->unauthorized_access_blocked();
         }
     }
 }
